API_Source_Code/Flask_APP_V1/requirements.txt
boto3==1.9.253
botocore==1.12.253
certifi==2019.9.11
chardet==3.0.4
Click==7.0
DateTime==4.3
docutils==0.15.2
Flask==1.1.1
git-lfs==1.5
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
MarkupSafe==1.1.1
numpy==1.17.2
pandas==0.25.1
psycopg2-binary==2.8.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
ta==0.4.5
urllib3==1.25.6
virtualenv==16.7.6
Werkzeug==0.16.0
zope.interface==4.6.0
